requirements.txt
xlrd
num2words
python-docx
PyYAML
scipy
google-api-python-client
pandas
google_auth_oauthlib
google_auth_httplib2
html2text
